soulswipe/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.0
python-dotenv==0.19.2
bcrypt==3.2.0
requests==2.28.1